:bug: automatic cluster formation using consul manual tests bug fixes

+++ internal/cluster/process.go
@@ -30,10 +30,6 @@ type procHandler struct {
 	httpHandler http.Handler
 }
 
-const (
-	automaticClusterFormationMaxRetry = 10
-)
-
 var _ proctree.ProcessHandler = (*procHandler)(nil)
 
 func (p *procHandler) Init(ctx actor.Context) proctree.ProcessResult {
@@ -55,11 +51,6 @@ func (p *procHandler) Init(ctx actor.Context) proctree.ProcessResult {
 
 		localEndpoint: localEndpoint,
 		endpoints:     make(map[string]*url.URL),
-	}
-
-	err = p.awaitAutoCluster()
-	if err != nil {
-		return proctree.Terminate(err)
 	}
 
 	if p.opts.ClusterJoinNode.JoinNodeID != "" && p.opts.ClusterJoinNode.JoinNodeEndpoint != nil {
@@ -87,11 +78,29 @@ func (p *procHandler) Init(ctx actor.Context) proctree.ProcessResult {
 		return proctree.Terminate(err)
 	}
 
+	/* If automatic cluster formation is enabled,
+	then ConsulService already discovered nodes and set them in ClusterNodeNode.
+	ConsulService starts before MagementServer therefore ManagementServer need not
+	wait here for ConsulService to finish discovering nodes */
 	if p.opts.ClusterJoinNode.JoinNodeID != "" && p.opts.ClusterJoinNode.JoinNodeEndpoint != nil {
 		joinAddr := fmt.Sprintf("%s/%s", p.opts.ClusterJoinNode.JoinNodeID, p.opts.ClusterJoinNode.JoinNodeEndpoint.Host)
 		_, err = p.mlist.Join([]string{joinAddr})
 		if err != nil {
-			return proctree.Terminate(err)
+			if !p.opts.AutomaticClusterFormation {
+				/* Terminate the process only when AutomaticClusterFormation is disabled
+				because otherwise the first node in the cluster will never know other nodes to connect to
+				therefore we cannot terminate the process for the very first node in the cluster */
+				logger.ErrorContext(ctx,
+					"memberlist join failed",
+					slog.Any("error", err),
+				)
+				return proctree.Terminate(err)
+			} else {
+				logger.WarnContext(ctx,
+					"memberlist join failed",
+					slog.Any("error", err),
+				)
+			}
 		}
 	}
 
@@ -117,30 +126,4 @@ func (p *procHandler) Terminate(ctx actor.Context, parentErr error) error {
 	}
 
 	return parentErr
-}
-
-// Wait for ConsulService to discover other nodes
-// ManagementServer and ConsulService share ClusterJoinNode.
-// ConsulService writes to ClusterJoinNode whereas ManagementServer reads ClusterJoinNode
-func (p *procHandler) awaitAutoCluster() error {
-	if !p.opts.AutomaticClusterFormation {
-		return nil
-	}
-
-	retryCount := 0
-	delay := 100 * time.Millisecond
-
-	for retryCount < automaticClusterFormationMaxRetry {
-		if p.opts.ClusterJoinNode.JoinNodeID == "" {
-			time.Sleep(delay)
-			// Exponential backoff
-			delay = delay * 2
-		}
-		retryCount++
-	}
-
-	if p.opts.ClusterJoinNode.JoinNodeID == "" {
-		return fmt.Errorf("exceed max retry to get join node through automatic cluster formation")
-	}
-	return nil
 }
